requirements.txt
APScheduler==3.11.0
discord.py==2.5.2
gspread==6.2.0
numpy==2.2.4
oauth2client==4.1.3
python-dotenv==1.1.0
Requests==2.32.3
beautifulsoup4==4.13.4